refactor: mock生成を別modに切り出す
