+++ src-tauri/src/infrastructure/windowsimpl/windows.rs
@@ -31,19 +31,6 @@ pub trait WindowsExt {
     fn proctail_manager(&self) -> &Self::ProcTailManager;
 }
 
-#[cfg(test)]
-mockall::mock! {
-    pub WindowsExtMock {}
-    impl WindowsExt for WindowsExtMock {
-        type ProcessWindows = crate::domain::windows::process::MockProcessWindows;
-        type ProcTail = crate::domain::windows::proctail::MockProcTail;
-        type ProcTailManager = crate::infrastructure::windowsimpl::proctail_manager::MockProcTailManagerTrait;
-
-        fn process(&self) -> &Self::ProcessWindows;
-        fn proctail(&self) -> &Self::ProcTail;
-        fn proctail_manager(&self) -> &Self::ProcTailManager;
-    }
-}
 
 impl WindowsExt for Windows {
     type ProcessWindows = WindowsImpl<Process>;
